feat: minor refactor (#6483)
