+++ frontend/src/views/Collection/hooks/useDragAndDrop/common/entities.ts
